sandbox: Split up LedgerApiServer.
CHANGELOG_BEGIN
CHANGELOG_END
