+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/LedgerApiServer.scala
@@ -3,37 +3,17 @@
 
 package com.daml.platform.apiserver
 
-import java.io.IOException
-import java.net.{BindException, InetAddress, InetSocketAddress}
-import java.util.UUID
-import java.util.concurrent.TimeUnit.SECONDS
-
 import akka.actor.ActorSystem
 import akka.stream.Materializer
 import com.codahale.metrics.MetricRegistry
-import com.daml.grpc.adapter.{AkkaExecutionSequencerPool, ExecutionSequencerFactory}
+import com.daml.grpc.adapter.ExecutionSequencerFactory
 import com.daml.logging.{ContextualizedLogger, LoggingContext}
 import com.daml.ports.Port
 import com.daml.resources.{Resource, ResourceOwner}
-import com.google.protobuf.Message
-import io.grpc.netty.NettyServerBuilder
-import io.grpc._
-import io.netty.channel.socket.nio.NioServerSocketChannel
-import io.netty.channel.{EventLoopGroup, ServerChannel}
+import io.grpc.ServerInterceptor
 import io.netty.handler.ssl.SslContext
 
 import scala.concurrent.{ExecutionContext, Future, Promise}
-import scala.util.control.NoStackTrace
-
-trait ApiServer {
-
-  /** the API port the server is listening on */
-  def port: Port
-
-  /** completes when all services have been closed during the shutdown */
-  def servicesClosed(): Future[Unit]
-
-}
 
 final class LedgerApiServer(
     createApiServices: (Materializer, ExecutionSequencerFactory) => Future[ApiServices],
@@ -73,7 +53,7 @@ final class LedgerApiServer(
         channelType,
         bossEventLoopGroup,
         workerEventLoopGroup,
-        apiServices,
+        apiServices.services,
       ).acquire()
       // Notify the caller that the services have been closed, so a reset request can complete
       // without blocking on the server terminating.
@@ -93,102 +73,4 @@ final class LedgerApiServer(
       }
     }
   }
-
-  private final class ExecutionSequencerFactoryOwner(implicit actorSystem: ActorSystem)
-      extends ResourceOwner[ExecutionSequencerFactory] {
-    // NOTE: Pick a unique pool name as we want to allow multiple LedgerApiServer instances,
-    // and it's pretty difficult to wait for the name to become available again.
-    // The name deregistration is asynchronous and the close method does not wait, and it isn't
-    // trivial to implement.
-    // https://doc.akka.io/docs/akka/2.5/actors.html#graceful-stop
-    private val poolName = s"ledger-api-server-rs-grpc-bridge-${UUID.randomUUID}"
-
-    private val ActorCount = Runtime.getRuntime.availableProcessors() * 8
-
-    override def acquire()(
-        implicit executionContext: ExecutionContext
-    ): Resource[ExecutionSequencerFactory] =
-      Resource(Future(new AkkaExecutionSequencerPool(poolName, ActorCount)))(_.closeAsync())
-  }
-
-  private final class GrpcServerOwner(
-      address: Option[String],
-      desiredPort: Port,
-      maxInboundMessageSize: Int,
-      sslContext: Option[SslContext] = None,
-      interceptors: List[ServerInterceptor] = List.empty,
-      metrics: MetricRegistry,
-      channelType: Class[_ <: ServerChannel],
-      bossEventLoopGroup: EventLoopGroup,
-      workerEventLoopGroup: EventLoopGroup,
-      apiServices: ApiServices,
-  ) extends ResourceOwner[Server] {
-    override def acquire()(implicit executionContext: ExecutionContext): Resource[Server] = {
-      val host = address.map(InetAddress.getByName).getOrElse(InetAddress.getLoopbackAddress)
-      Resource(Future {
-        val builder = NettyServerBuilder.forAddress(new InetSocketAddress(host, desiredPort.value))
-        builder.sslContext(sslContext.orNull)
-        builder.channelType(classOf[NioServerSocketChannel])
-        builder.permitKeepAliveTime(10, SECONDS)
-        builder.permitKeepAliveWithoutCalls(true)
-        builder.directExecutor()
-        builder.maxInboundMessageSize(maxInboundMessageSize)
-        interceptors.foreach(builder.intercept)
-        builder.intercept(new MetricsInterceptor(metrics))
-        builder.channelType(channelType)
-        builder.bossEventLoopGroup(bossEventLoopGroup)
-        builder.workerEventLoopGroup(workerEventLoopGroup)
-        apiServices.services.foreach { service =>
-          builder.addService(service)
-          toLegacyService(service).foreach(builder.addService)
-        }
-        val server = builder.build()
-        try {
-          server.start()
-        } catch {
-          case e: IOException if e.getCause != null && e.getCause.isInstanceOf[BindException] =>
-            throw new UnableToBind(desiredPort, e.getCause)
-        }
-        server
-      })(server => Future(server.shutdown().awaitTermination()))
-    }
-
-    // This exposes the existing services under com.daml also under com.digitalasset.
-    // This is necessary to allow applications built with an earlier version of the SDK
-    // to still work.
-    // The "proxy" services will not show up on the reflection service, because of the way it
-    // processes service definitions via protobuf file descriptors.
-    private def toLegacyService(service: BindableService): Option[ServerServiceDefinition] = {
-      val `com.daml` = "com.daml"
-      val `com.digitalasset` = "com.digitalasset"
-
-      val damlDef = service.bindService()
-      val damlDesc = damlDef.getServiceDescriptor
-      // Only add "proxy" services if it actually contains com.daml in the service name.
-      // There are other services registered like the reflection service, that doesn't need the special treatment.
-      if (damlDesc.getName.contains(`com.daml`)) {
-        val digitalassetName = damlDesc.getName.replace(`com.daml`, `com.digitalasset`)
-        val digitalassetDef = ServerServiceDefinition.builder(digitalassetName)
-        damlDef.getMethods.forEach { methodDef =>
-          val damlMethodDesc = methodDef.getMethodDescriptor
-          val digitalassetMethodName =
-            damlMethodDesc.getFullMethodName.replace(`com.daml`, `com.digitalasset`)
-          val digitalassetMethodDesc =
-            damlMethodDesc.toBuilder.setFullMethodName(digitalassetMethodName).build()
-          val _ = digitalassetDef.addMethod(
-            digitalassetMethodDesc.asInstanceOf[MethodDescriptor[Message, Message]],
-            methodDef.getServerCallHandler.asInstanceOf[ServerCallHandler[Message, Message]]
-          )
-        }
-        Option(digitalassetDef.build())
-      } else None
-    }
-  }
-
-  final class UnableToBind(port: Port, cause: Throwable)
-      extends RuntimeException(
-        s"LedgerApiServer was unable to bind to port $port. " +
-          "User should terminate the process occupying the port, or choose a different one.",
-        cause)
-      with NoStackTrace
 }
